% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_budswelling.R
\docType{data}
\name{budswelling}
\alias{budswelling}
\title{Regional budswelling data}
\format{
netCDF
}
\description{
A netCDF containing the DOY (day of year) for budswelling in the year 2023
provided by the Agricultural meteorology Department of the DWD.
}
\details{
The dataset can be loaded using load_test("budswelling").

Spatial resolution: 1 km x 1 km
Projection: DHDN / 3-degree Gauss-Kruger zone 3 (EPSG:31467)
Parameter: DOY for first day of budswelling
}
\keyword{datasets}
