% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_baskerville.R
\name{calc_baskerville}
\alias{calc_baskerville}
\title{Calculate degree days using the single sine method of Baskerville}
\usage{
calc_baskerville(tmin, tmax, ldt)
}
\arguments{
\item{tmin}{min temperature - numeric}

\item{tmax}{max temperature - numeric}

\item{ldt}{lower development threshold aka base temperature - numeric}
}
\value{
Degree days calculated by using the single sine method of Baskerville.
}
\description{
The function is vectorized.
To use it with SpatRaster data see ?terra::lapp

More Information about the Method itself:
Baskerville, G.L. and P. Emin. 1969. Rapid estimation of heat accumulation from maximum and minimum temperature. Ecology 50:514-517.
}
\seealso{
Other Temperature sum methods: 
\code{\link{calc_tsum}()}
}
\concept{Temperature sum methods}
\keyword{internal}
