% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_degreedays.R
\name{calc_degreedays}
\alias{calc_degreedays}
\title{Calculate degree days}
\usage{
calc_degreedays(x, ts_start, ts_end, ldt, method)
}
\arguments{
\item{x}{SpatRaster list (tmean, tmax, tmin) - numeric - with time attribute}

\item{ts_start}{start of timeseries - Date}

\item{ts_end}{end of timeseries - Date}

\item{ldt}{lower development threshold - numeric}

\item{method}{name of degreedays/temperature sum method - character - either
"baskerville" or "tsum"}
}
\value{
SpatRaster of growing (summed up) degree days.
}
\description{
Calculate degree days
}
\seealso{
Other Models: 
\code{\link{calc_mortality}()},
\code{\link{calc_phenology}()},
\code{\link{calc_sumefftemp}()}
}
\concept{Models}
\keyword{internal}
