% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_first.R
\name{calc_first}
\alias{calc_first}
\title{Calculate first TRUE}
\usage{
calc_first(...)
}
\arguments{
\item{...}{SpatRasters/SpatRaster list - logical - with time attribute}
}
\value{
SpatRaster - logical - with time attribute
}
\description{
Takes multiple logical SpatRasters with time attribute and returns SpatRaster with common time and TRUE if any SpatRaster is TRUE.
}
\seealso{
Other Calculation: 
\code{\link{calc_last}()},
\code{\link{calc_mean}()}
}
\concept{Calculation}
\keyword{internal}
