% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_last.R
\name{calc_last}
\alias{calc_last}
\title{Calculate last TRUE}
\usage{
calc_last(x)
}
\arguments{
\item{x}{SpatRaster - logical - with time attribute}
}
\value{
SpatRaster with time serial number (origin = lubridate::origin) of
last TRUE layer.
}
\description{
Takes multiple logical SpatRasters with time attribute and
returns SpatRaster with time serial number of last TRUE layer time.
}
\seealso{
Other Calculation: 
\code{\link{calc_first}()},
\code{\link{calc_mean}()}
}
\concept{Calculation}
\keyword{internal}
