% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mean.R
\name{calc_mean}
\alias{calc_mean}
\title{Calculate mean TRUE}
\usage{
calc_mean(x)
}
\arguments{
\item{x}{SpatRaster list - logical - with time attribute}
}
\value{
SpatRaster - logical - with time attribute
}
\description{
Takes multiple logical SpatRasters with time attribute and
returns SpatRaster with the mean time serial number of first and last TRUE
value.
}
\seealso{
Other Calculation: 
\code{\link{calc_first}()},
\code{\link{calc_last}()}
}
\concept{Calculation}
\keyword{internal}
