% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_mortality.R
\name{calc_mortality}
\alias{calc_mortality}
\title{Calculate mortality}
\usage{
calc_mortality(x, budswelling, hatch, mot = 0)
}
\arguments{
\item{x}{SpatRaster list / dataframe with temperature data and time
attribute/column}

\item{budswelling}{SpatRaster - logical - T/F budswelling/no_budswelling}

\item{hatch}{SpatRaster - logical - T/F hatch/no_hatch}

\item{mot}{mortality threshold - numeric - is used as lower development
threshold calculating degree days}
}
\value{
SpatRaster of starving related mortality in \%.
}
\description{
Calculate mortality
}
\seealso{
Other Models: 
\code{\link{calc_degreedays}()},
\code{\link{calc_phenology}()},
\code{\link{calc_sumefftemp}()}
}
\concept{Models}
\keyword{internal}
