% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_phenology.R
\name{calc_phenology}
\alias{calc_phenology}
\title{Calculate phenological models}
\usage{
calc_phenology(x, params)
}
\arguments{
\item{x}{SpatRaster list - with tmean, tmin, tmax and time attribute}

\item{params}{parameter list}
}
\value{
SpatRaster - logical - with event occurred/not TRUE/FALSE
}
\description{
This function is checking the dependencies of the model and dependent on them:

1 Calculate the needed sum of effective (SET) temperatures for the phenological event to happen
2 Calculate the summed degree days (depending on hatch or not)
3 Check if the summed degree days have reached the needed sum of effective temperatures
}
\seealso{
Other Models: 
\code{\link{calc_degreedays}()},
\code{\link{calc_mortality}()},
\code{\link{calc_sumefftemp}()}
}
\concept{Models}
\keyword{internal}
