% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tsum.R
\name{calc_tsum}
\alias{calc_tsum}
\title{Calculate temperature sum}
\usage{
calc_tsum(tmean, ldt)
}
\arguments{
\item{tmean}{numeric - mean temperature}

\item{ldt}{numeric - lower development threshold aka base temperature.}
}
\value{
The temperature sum over the lower development threshold.
}
\description{
The function is vectorized.
To use it with SpatRaster-Data see ?terra::lapp.
}
\seealso{
Other Temperature sum methods: 
\code{\link{calc_baskerville}()}
}
\concept{Temperature sum methods}
\keyword{internal}
