% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_function.R
\name{call_function}
\alias{call_function}
\title{Call a function from PHENTHAUproc package}
\usage{
call_function(f, data, params)
}
\arguments{
\item{f}{function}

\item{data}{SpatRaster list (tmean, tmax, tmin) - numeric - with time
attribute}

\item{params}{parameter list}
}
\description{
calls a function from PHENTHAUproc package
}
\seealso{
Other Helper: 
\code{\link{check_data_with_params}()},
\code{\link{check_dimension_and_time}()},
\code{\link{convert_dwd_to_phenthau}()},
\code{\link{get_date}()},
\code{\link{get_formalArgs}()},
\code{\link{get_time}()},
\code{\link{load_test}()},
\code{\link{remove_false}()},
\code{\link{set_attributes}()},
\code{\link{subset_time}()},
\code{\link{timename}()}
}
\concept{Helper}
\keyword{internal}
