% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_with_params.R
\name{check_data_with_params}
\alias{check_data_with_params}
\title{Check data for model requirements}
\usage{
check_data_with_params(x, params)
}
\arguments{
\item{x}{SpatRaster list - numeric - with time attribute}

\item{params}{parameter list}
}
\value{
No return value, called for side effects.
}
\description{
Check data for model requirements
}
\seealso{
Other Helper: 
\code{\link{call_function}()},
\code{\link{check_dimension_and_time}()},
\code{\link{convert_dwd_to_phenthau}()},
\code{\link{get_date}()},
\code{\link{get_formalArgs}()},
\code{\link{get_time}()},
\code{\link{load_test}()},
\code{\link{remove_false}()},
\code{\link{set_attributes}()},
\code{\link{subset_time}()},
\code{\link{timename}()}
}
\concept{Helper}
\keyword{internal}
