% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dimension_and_time.R
\name{check_dimension_and_time}
\alias{check_dimension_and_time}
\title{Check dimension and time attribute of a SpatRaster list}
\usage{
check_dimension_and_time(...)
}
\arguments{
\item{...}{SpatRaster/SpatRaster list/data.frame/vector}
}
\value{
No return value, called for side effects.
}
\description{
Check dimension and time attribute of a SpatRaster list
}
\seealso{
Other Helper: 
\code{\link{call_function}()},
\code{\link{check_data_with_params}()},
\code{\link{convert_dwd_to_phenthau}()},
\code{\link{get_date}()},
\code{\link{get_formalArgs}()},
\code{\link{get_time}()},
\code{\link{load_test}()},
\code{\link{remove_false}()},
\code{\link{set_attributes}()},
\code{\link{subset_time}()},
\code{\link{timename}()}
}
\concept{Helper}
\keyword{internal}
