% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_doy_to_logical.R
\name{convert_doy_to_logical}
\alias{convert_doy_to_logical}
\title{Convert day of year (doy) to logical}
\usage{
convert_doy_to_logical(x, from, to, by = "days")
}
\arguments{
\item{x}{SpatRaster - numeric - value is day of year}

\item{from}{Date - ymd - first day of time attribute}

\item{to}{Date - ymd - last day of time attribute}

\item{by}{character - either "days", "weeks" or "months". Reduces return to one day per day/week/month}
}
\value{
SpatRaster - logical - TRUE/FALSE before_doy/equal_after_doy
}
\description{
Convert day of year (doy) to logical
}
\seealso{
Other SpatRaster transformation: 
\code{\link{convert_df_to_srl}()},
\code{\link{convert_hour_to_meanminmax}()},
\code{\link{convert_logical_to_doy}()},
\code{\link{convert_logical_to_time}()},
\code{\link{convert_sr_to_cvec}()}
}
\concept{SpatRaster transformation}
\keyword{internal}
