% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_logical_to_doy.R
\name{convert_logical_to_doy}
\alias{convert_logical_to_doy}
\title{Convert logical to day of year (doy)}
\usage{
convert_logical_to_doy(x, from = NULL, to = NULL)
}
\arguments{
\item{x}{SpatRaster - logical - with time attribute}

\item{from}{\itemize{
\item Date - ymd - only if SpatRaster should be subset before being
transformed.
}}

\item{to}{\itemize{
\item Date - ymd - only if SpatRaster should be subset before being
transformed.
}}
}
\value{
SpatRaster - numeric - day of year
}
\description{
Converts logical SpatRaster with time attribute and one layer
per day to single layer with the day of the year (doy) where the layer is
1/TRUE first.
}
\seealso{
Other SpatRaster transformation: 
\code{\link{convert_df_to_srl}()},
\code{\link{convert_doy_to_logical}()},
\code{\link{convert_hour_to_meanminmax}()},
\code{\link{convert_logical_to_time}()},
\code{\link{convert_sr_to_cvec}()}
}
\concept{SpatRaster transformation}
\keyword{internal}
