% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_logical_to_time.R
\name{convert_logical_to_time}
\alias{convert_logical_to_time}
\title{Convert logical to time}
\usage{
convert_logical_to_time(x)
}
\arguments{
\item{x}{SpatRaster - logical - with time attribute}
}
\value{
SpatRaster - numeric
}
\description{
Converts logical SpatRasters to single layer SpatRaster with timeserialnumber as value for first TRUE
}
\seealso{
Other SpatRaster transformation: 
\code{\link{convert_df_to_srl}()},
\code{\link{convert_doy_to_logical}()},
\code{\link{convert_hour_to_meanminmax}()},
\code{\link{convert_logical_to_doy}()},
\code{\link{convert_sr_to_cvec}()}
}
\concept{SpatRaster transformation}
\keyword{internal}
