% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_date.R
\name{get_date}
\alias{get_date}
\title{Get Date format out of numeric year and monthday character value}
\usage{
get_date(year, monthday, prevyear = FALSE)
}
\arguments{
\item{year}{numeric year}

\item{monthday}{character with month and day i.e. ("-02-01" or "Feb 01")}

\item{prevyear}{if TRUE starts with previous year}
}
\description{
Get Date format out of numeric year and monthday character value
}
\seealso{
Other Helper: 
\code{\link{call_function}()},
\code{\link{check_data_with_params}()},
\code{\link{check_dimension_and_time}()},
\code{\link{convert_dwd_to_phenthau}()},
\code{\link{get_formalArgs}()},
\code{\link{get_time}()},
\code{\link{load_test}()},
\code{\link{remove_false}()},
\code{\link{set_attributes}()},
\code{\link{subset_time}()},
\code{\link{timename}()}
}
\concept{Helper}
\keyword{internal}
