% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_formalArgs.R
\name{get_formalArgs}
\alias{get_formalArgs}
\title{Get formalArgs of a function from a list of parameter}
\usage{
get_formalArgs(p, f)
}
\arguments{
\item{p}{parameter list}

\item{f}{function}
}
\description{
Get formalArgs of a function from a list of parameter
}
\seealso{
Other Helper: 
\code{\link{call_function}()},
\code{\link{check_data_with_params}()},
\code{\link{check_dimension_and_time}()},
\code{\link{convert_dwd_to_phenthau}()},
\code{\link{get_date}()},
\code{\link{get_time}()},
\code{\link{load_test}()},
\code{\link{remove_false}()},
\code{\link{set_attributes}()},
\code{\link{subset_time}()},
\code{\link{timename}()}
}
\concept{Helper}
\keyword{internal}
