% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_legend.R
\name{get_legend}
\alias{get_legend}
\title{Get legend for PHENTHAUproc models}
\usage{
get_legend(x = "stages")
}
\arguments{
\item{x}{legend name - character - Available legends: "stages", "mortality",
"ppa_biocide"}
}
\value{
dataframe with ID, category and colors
}
\description{
Get legend for PHENTHAUproc models
}
\examples{
# return legend for development stages
get_legend("stages")

}
\seealso{
Other Main: 
\code{\link{mortality}()},
\code{\link{parameter}()},
\code{\link{phenology}()},
\code{\link{phenthau}()}
}
\concept{Main}
