% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time.R
\name{get_time}
\alias{get_time}
\title{Get time attribute}
\usage{
get_time(x)
}
\arguments{
\item{x}{SpatRaster, data.frame, list of SpatRaster or SpatRasterDataset}
}
\value{
If x is a data.frame returns the time column, else it returns the time attribute of a SpatRaster or the shared time attribute of multiple SpatRaster.
}
\description{
Get time attribute
}
\seealso{
Other Helper: 
\code{\link{call_function}()},
\code{\link{check_data_with_params}()},
\code{\link{check_dimension_and_time}()},
\code{\link{convert_dwd_to_phenthau}()},
\code{\link{get_date}()},
\code{\link{get_formalArgs}()},
\code{\link{load_test}()},
\code{\link{remove_false}()},
\code{\link{set_attributes}()},
\code{\link{subset_time}()},
\code{\link{timename}()}
}
\concept{Helper}
\keyword{internal}
