% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_test.R
\name{load_test}
\alias{load_test}
\title{Load test data}
\usage{
load_test(type = "SpatRaster")
}
\arguments{
\item{type}{character, either day/hour/SpatRaster/SpatRaster_hour/budswelling}
}
\value{
data.frame/SpatRaster list with test data
}
\description{
Load test data
}
\seealso{
Other Helper: 
\code{\link{call_function}()},
\code{\link{check_data_with_params}()},
\code{\link{check_dimension_and_time}()},
\code{\link{convert_dwd_to_phenthau}()},
\code{\link{get_date}()},
\code{\link{get_formalArgs}()},
\code{\link{get_time}()},
\code{\link{remove_false}()},
\code{\link{set_attributes}()},
\code{\link{subset_time}()},
\code{\link{timename}()}
}
\concept{Helper}
