% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_test.R
\name{load_test_bs}
\alias{load_test_bs}
\title{load SpatRaster - budswelling - day of year - 16 Rasterpoints around FVA 2023}
\usage{
load_test_bs()
}
\description{
load SpatRaster - budswelling - day of year - 16 Rasterpoints around FVA 2023
}
\keyword{internal}
