% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_test.R
\name{load_test_day}
\alias{load_test_day}
\title{load dataframe - daily tmean tmin tmax - Freiburg 2020}
\usage{
load_test_day()
}
\description{
load dataframe - daily tmean tmin tmax - Freiburg 2020
}
\keyword{internal}
