% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_test.R
\name{load_test_srl}
\alias{load_test_srl}
\title{load SpatRaster list - daily - 16 Rasterpoints around FVA 2020}
\usage{
load_test_srl()
}
\description{
load SpatRaster list - daily - 16 Rasterpoints around FVA 2020
}
\keyword{internal}
