% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_test.R
\name{load_test_srl_hour}
\alias{load_test_srl_hour}
\title{load SpatRaster - hourly - 16 Rasterpoints around FVA 2022 - 2023}
\usage{
load_test_srl_hour()
}
\description{
load SpatRaster - hourly - 16 Rasterpoints around FVA 2022 - 2023
}
\keyword{internal}
