% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_station.R
\name{plot_station}
\alias{plot_station}
\title{Plot local PHENTHAUproc in a circle plot}
\usage{
plot_station(x, main = "PHENTHAUproc")
}
\arguments{
\item{x}{output of phenthau - dataframe - local PHENTHAUproc}

\item{main}{title - character}
}
\value{
A plot showing local PHENTHAUproc results in a circle plot presentation.
}
\description{
Plot local PHENTHAUproc in a circle plot
}
\examples{

# load daily test data from Freiburg
fr_df <- load_test("day")
fr <- phenthau(fr_df)

plot_station(fr)
}
\references{
Allan Cameron 2022: Allan Cameron (2022), Answer to the question "Is it possible to create a circular timeline plot with ggplot?" on Stackoverflow, (\url{https://stackoverflow.com/questions/71023419/is-it-possible-to-create-a-circular-timeline-plot-with-ggplot})
}
\seealso{
Other Plot: 
\code{\link{plot_date}()},
\code{\link{plot_stages}()},
\code{\link{plot_station_step}()}
}
\concept{Plot}
