% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_station_step.R
\name{plot_station_step}
\alias{plot_station_step}
\title{Plot local PHENTHAUproc in a step plot}
\usage{
plot_station_step(x)
}
\arguments{
\item{x}{output of phenthau - dataframe - local PHENTHAUproc}
}
\value{
A plot showing local 'PHENTHAUproc' results in a stair step presentation.
}
\description{
Plots the development stages of OPM.
}
\examples{

fr_df <- load_test("day")
fr <- phenthau(fr_df)

plot_station_step(fr)
}
\seealso{
Other Plot: 
\code{\link{plot_date}()},
\code{\link{plot_stages}()},
\code{\link{plot_station}()}
}
\concept{Plot}
