% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_regional_hour.R
\docType{data}
\name{regional_hour}
\alias{regional_hour}
\title{Regional Weather data}
\format{
netCDF
}
\description{
A dataset containing hourly mean temperatures in °C.
The dataset is a 4*4 pixel cutout centered at FVA provided by the
Agricultural meteorology Department of the DWD.
}
\details{
The dataset can be loaded using load_test("SpatRaster_hour").

Spatial resolution: 1 km x 1 km
Projection: DHDN / 3-degree Gauss-Kruger zone 3 (EPSG:31467)
Parameter: air temperature at 2 m

The dataset is a list with three SpatRaster as objects.

The list objects are:

\itemize{
\item fva_hour: SpatRaster with hourly mean temperature in °C - numeric
}
}
\keyword{datasets}
