% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_false.R
\name{remove_false}
\alias{remove_false}
\title{Remove FALSE}
\usage{
remove_false(x, y)
}
\arguments{
\item{x}{SpatRaster/SpatRaster list - numeric - with time attribute}

\item{y}{SpatRaster - logical - with time attribute}
}
\description{
Set values in SpatRaster or SpatRaster list to -999 if y is FALSE
}
\seealso{
Other Helper: 
\code{\link{call_function}()},
\code{\link{check_data_with_params}()},
\code{\link{check_dimension_and_time}()},
\code{\link{convert_dwd_to_phenthau}()},
\code{\link{get_date}()},
\code{\link{get_formalArgs}()},
\code{\link{get_time}()},
\code{\link{load_test}()},
\code{\link{set_attributes}()},
\code{\link{subset_time}()},
\code{\link{timename}()}
}
\concept{Helper}
\keyword{internal}
