% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_attributes.R
\name{set_attributes}
\alias{set_attributes}
\title{Set attributes for SpatRaster}
\usage{
set_attributes(x, type)
}
\arguments{
\item{x}{SpatRaster - to assign attributes to}

\item{type}{character - type of attributes i.e. "stages"}
}
\description{
Set attributes for SpatRaster
}
\seealso{
Other Helper: 
\code{\link{call_function}()},
\code{\link{check_data_with_params}()},
\code{\link{check_dimension_and_time}()},
\code{\link{convert_dwd_to_phenthau}()},
\code{\link{get_date}()},
\code{\link{get_formalArgs}()},
\code{\link{get_time}()},
\code{\link{load_test}()},
\code{\link{remove_false}()},
\code{\link{subset_time}()},
\code{\link{timename}()}
}
\concept{Helper}
\keyword{internal}
