% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_time.R
\name{subset_time}
\alias{subset_time}
\title{Subset data.frame/SpatRaster/List within from and end time}
\usage{
subset_time(x, from, to)
}
\arguments{
\item{x}{data.frame with time column, SpatRaster with time attribute or List
of SpatRasters with same time attributes}

\item{from}{first day of subset}

\item{to}{last day of subset}
}
\description{
Subset data.frame/SpatRaster/List within from and end time
}
\seealso{
Other Helper: 
\code{\link{call_function}()},
\code{\link{check_data_with_params}()},
\code{\link{check_dimension_and_time}()},
\code{\link{convert_dwd_to_phenthau}()},
\code{\link{get_date}()},
\code{\link{get_formalArgs}()},
\code{\link{get_time}()},
\code{\link{load_test}()},
\code{\link{remove_false}()},
\code{\link{set_attributes}()},
\code{\link{timename}()}
}
\concept{Helper}
\keyword{internal}
