% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timename.R
\name{timename}
\alias{timename}
\title{Create a name with time stamp}
\usage{
timename(x, model, desc = NULL)
}
\arguments{
\item{x}{SpatRaster with time attribute}

\item{model}{character - name of model}

\item{desc}{character - additional description (optional)}
}
\description{
Create a name with time stamp
}
\seealso{
Other Helper: 
\code{\link{call_function}()},
\code{\link{check_data_with_params}()},
\code{\link{check_dimension_and_time}()},
\code{\link{convert_dwd_to_phenthau}()},
\code{\link{get_date}()},
\code{\link{get_formalArgs}()},
\code{\link{get_time}()},
\code{\link{load_test}()},
\code{\link{remove_false}()},
\code{\link{set_attributes}()},
\code{\link{subset_time}()}
}
\concept{Helper}
\keyword{internal}
