% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dwd_to_phenthau.R
\name{convert_dwd_to_phenthau}
\alias{convert_dwd_to_phenthau}
\title{Convert hourly DWD temperature data to PHENTHAUproc input}
\usage{
convert_dwd_to_phenthau(x)
}
\arguments{
\item{x}{filepath to unzipped DWD temperature data (text file)}
}
\value{
A dataframe with date, hour and mean air temperature (tmean).
}
\description{
Convert hourly DWD temperature data to PHENTHAUproc input
}
\seealso{
Other Helper: 
\code{\link{call_function}()},
\code{\link{check_data_with_params}()},
\code{\link{check_dimension_and_time}()},
\code{\link{get_date}()},
\code{\link{get_formalArgs}()},
\code{\link{get_time}()},
\code{\link{load_test}()},
\code{\link{remove_false}()},
\code{\link{set_attributes}()},
\code{\link{subset_time}()},
\code{\link{timename}()}
}
\concept{Helper}
