% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stages.R
\name{plot_stages}
\alias{plot_stages}
\title{Plot regional 'PHENTHAUproc' - Stages}
\usage{
plot_stages(x, time = NULL, ...)
}
\arguments{
\item{x}{SpatRaster stages output of phenthau - numeric}

\item{time}{day to plot - Date or ymd character}

\item{...}{arguments passed along to terra::plot}
}
\value{
A plot of the 'PHENTHAUproc' stages of the last/chosen time with preset levels and colors.
}
\description{
wrapper to plot stages of OPM with assigned names and colors
}
\details{
phenthau returns a list of SpatRasters. The stages object is describing the development stages of oak processionary moth.
This function plots the stages SpatRaster with the right names and colors. to see the assignment check get_legend("stages")
}
\seealso{
Other Plot: 
\code{\link{plot_date}()},
\code{\link{plot_station_step}()},
\code{\link{plot_station}()}
}
\concept{Plot}
